+++ PayPal/src/test/java/com/braintreepayments/api/PayPalVaultRequestUnitTest.java
@@ -1,12 +1,10 @@
 package com.braintreepayments.api;
 
-import android.content.pm.Signature;
 import android.os.Parcel;
 
 import org.json.JSONException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mockito;
 import org.robolectric.RobolectricTestRunner;
 
 import java.util.ArrayList;
@@ -31,7 +29,7 @@ public class PayPalVaultRequestUnitTest {
         assertNull(request.getLandingPageType());
         assertFalse(request.getShouldOfferCredit());
         assertFalse(request.hasUserLocationConsent());
-        assertFalse(request.isUseAppLinkReturn());
+        assertFalse(request.isAppLinkEnabled());
     }
 
     @Test
@@ -52,7 +50,7 @@ public class PayPalVaultRequestUnitTest {
         request.setRiskCorrelationId("123-correlation");
         request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
         request.setShouldOfferCredit(true);
-        request.setUseAppLinkReturn(true);
+        request.setAppLinkEnabled(true);
 
         assertEquals("US", request.getLocaleCode());
         assertEquals("Billing Agreement Description", request.getBillingAgreementDescription());
@@ -63,7 +61,7 @@ public class PayPalVaultRequestUnitTest {
         assertEquals(PayPalRequest.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
         assertTrue(request.getShouldOfferCredit());
         assertTrue(request.hasUserLocationConsent());
-        assertTrue(request.isUseAppLinkReturn());
+        assertTrue(request.isAppLinkEnabled());
     }
 
     @Test
@@ -74,7 +72,7 @@ public class PayPalVaultRequestUnitTest {
         request.setShippingAddressRequired(true);
         request.setShippingAddressEditable(true);
         request.setShouldOfferCredit(true);
-        request.setUseAppLinkReturn(true);
+        request.setAppLinkEnabled(true);
 
         PostalAddress postalAddress = new PostalAddress();
         postalAddress.setRecipientName("Postal Address");
@@ -109,7 +107,7 @@ public class PayPalVaultRequestUnitTest {
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
         assertTrue(result.hasUserLocationConsent());
-        assertTrue(result.isUseAppLinkReturn());
+        assertTrue(result.isAppLinkEnabled());
     }
 
     @Test