+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Venmo
   * Send `link_type` in `event_params` to PayPal's analytics service (FPTI)
 * PayPal
-  * Add `useAppLinkReturn` property to `PayPalRequest` for Android App Link support
+  * Add `appLinkEnabled` property to `PayPalRequest` for Android App Link support
   
 ## 4.45.0 (2024-04-16)